refactor(date): add types folder
