+++ packages/date/package.json
@@ -3,7 +3,7 @@
   "version": "2.1.10",
   "description": "Wrapper for react-dates to work with formik",
   "main": "index.js",
-  "typings": "typings/index.d.ts",
+  "types": "index.d.ts",
   "keywords": [
     "availity",
     "reactstrap",
@@ -23,7 +23,6 @@
   "devDependencies": {
     "@availity/form": "^0.7.0",
     "@availity/yup": "^3.0.11",
-    "@types/react-dates": "^21.8.2",
     "formik": "^2.2.8",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
@@ -31,7 +30,7 @@
     "yup": "^0.30.0"
   },
   "peerDependencies": {
-    "@availity/form": "^0.1.0 || ^0.2.0 || ^0.3.0 || ^0.4.0  || ^0.5.0",
+    "@availity/form": "^0.1.0 || ^0.2.0 || ^0.3.0 || ^0.4.0  || ^0.5.0 || ^0.6.0 || ^0.7.0",
     "@availity/yup": "^3.0.2",
     "formik": "^2.0.1",
     "react": ">=16.3.0",
@@ -40,6 +39,7 @@
   },
   "dependencies": {
     "@availity/icon": "^0.8.0",
+    "@types/react-dates": "^21.8.2",
     "classnames": "^2.3.1",
     "lodash": "^4.17.21",
     "moment": "^2.29.1",
