+++ packages/date/types/DateField.d.ts
@@ -1,13 +1,13 @@
 import * as React from 'react';
 import { DateProps } from './Date';
 
-export interface DateFieldProps extends DateProps {
+export type DateFieldProps = {
   id: string;
   label?: React.ReactNode;
   labelClass?: string;
   labelHidden?: boolean;
   labelAttrs?: React.HTMLAttributes<HTMLLabelElement>;
-}
+} & DateProps;
 
 declare class DateField extends React.Component<DateFieldProps> {}
